@@ -1,11 +1,5 @@
 /*----------------------------------------------------------------------------*/
 /* Source File:   MESSAGECONTROLLERINTEGRATIONTEST.JAVA                       */
-/* Description:   REST Api for ChuckNorrisData and ChuckNorrisWrapperData     */
-/*                end-points.                                                 */
-/* Author:        Carlos Adolfo Ortiz Quirós (COQ)                            */
-/* Date:          Oct.15/2019                                                 */
-/* Last Modified: Jul.05/2020                                                 */
-/* Version:       1.1                                                         */
 /* Copyright (c), 2019, 2020 CSoftZ                                           */
 /*----------------------------------------------------------------------------*/
 /*-----------------------------------------------------------------------------
@@ -18,13 +12,13 @@ import static org.springframework.restdocs.operation.preprocess.Preprocessors.pr
 import static org.springframework.restdocs.operation.preprocess.Preprocessors.prettyPrint;
 import static org.springframework.restdocs.webtestclient.WebTestClientRestDocumentation.document;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
 import org.springframework.boot.test.autoconfigure.web.reactive.WebFluxTest;
 import org.springframework.context.annotation.Import;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.reactive.server.WebTestClient;
 
 import com.example.chuck.norris.wrapper.api.controller.ChuckNorrisController;
@@ -35,14 +29,14 @@ import com.example.chuck.norris.wrapper.service.ChuckNorrisJokesService;
  * <p>As this is an integration test all of external services must be injected. That is why the @Import is there.</p>
  *
  * @author Carlos Adolfo Ortiz Quirós (COQ)
- * @version 1.1, Jul.05/2020
+ * @version 1.1, Dec.12/2020
  * @since 11 (JDK), Oct.15/2019
  */
-@RunWith(SpringRunner.class)
+@ExtendWith(SpringExtension.class)
 @AutoConfigureRestDocs(outputDir = "target/snippets", uriScheme = "http", uriHost = "api.example.com")
 @WebFluxTest(ChuckNorrisController.class)
 @Import(ChuckNorrisJokesService.class)
-public class ChuckNorrisControllerIntegrationTest {
+class ChuckNorrisControllerIntegrationTest {
 
     @Autowired
     private WebTestClient webTestClient;
@@ -52,7 +46,7 @@ public class ChuckNorrisControllerIntegrationTest {
      * Spring WebFlux testing capabilities.
      */
     @Test
-    public void shouldReturnChuckNorrisData() {
+    void shouldReturnChuckNorrisData() {
         this.webTestClient
             .get()
             .uri("/api/v1/chuck/norris/random/string")
